+++ tests/testthat/test-error-checks.R
@@ -145,6 +145,23 @@ test_that("isEmpty returns FALSE when objects are not empty", {
   expect_false(isEmpty(""))
 })
 
+test_that("hasEmptyString returns TRUE when strings are empty", {
+  expect_true(hasEmptyString(character(0)))
+  expect_true(hasEmptyString(c("", "y")))
+  expect_true(hasEmptyString(list("", "y")))
+  expect_true(hasEmptyString(c("", NA)))
+  expect_true(hasEmptyString(list("", NA)))
+  expect_true(hasEmptyString(NA))
+  expect_true(hasEmptyString(c(NA, "x", "y")))
+  expect_true(hasEmptyString(list(NA, "x", "y")))
+})
+
+test_that("hasEmptyString returns FALSE when string are not empty", {
+  expect_false(hasEmptyString("   abc   "))
+  expect_false(hasEmptyString(c("x", "y")))
+  expect_false(hasEmptyString(list("x", "y")))
+})
+
 test_that("hasOnlyDistinctValues returns TRUE if values are distinct", {
   expect_true(hasOnlyDistinctValues(c("x", NA, "y")))
   expect_true(hasOnlyDistinctValues(list("x", NA, "y")))